LPS-95113 Check before deletion whether the Layout has already been deleted as a child of another Layout

+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/exportimport/data/handler/StagedLayoutSetStagedModelDataHandler.java
@@ -138,7 +138,11 @@ public class StagedLayoutSetStagedModelDataHandler
 				layout.getSourcePrototypeLayoutUuid(),
 				layoutSetPrototype.getGroupId(), true);
 
-			if (sourcePrototypeLayout == null) {
+			if ((sourcePrototypeLayout == null) &&
+				_layoutLocalService.hasLayout(
+					layout.getUuid(), layout.getGroupId(),
+					layout.isPrivateLayout())) {
+
 				_layoutLocalService.deleteLayout(
 					layout, false,
 					ServiceContextThreadLocal.getServiceContext());
